Update Tor exit nodes source and improve IP filtering
The source of the Tor exit nodes list has been changed to use torproject.org for IPv4 and openinternet.io for IPv6 due to the fact that torproject.org does not provide IPv6 list. In addition, added logic to filter out IPv4 addresses and comments from IPv6 list. This ensures accuracy and reliability of the lists being obtained.